pyproject.toml
[tool.poetry]
name = "temporalloop"
version = "0.2.0"
description = ""
authors = ["ant31 <2t.antoine@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
tempoloop = 'temporalloop.main:main'

[tool.poetry.dependencies]
python = ">=3.11,<4"
temporalio = "^1.1"
asyncio = "^3.4.3"
pydantic = "^2.1"
sentry-sdk = "^1"
click = "^8.1.3"
pyyaml = "^6.0"
pydantic-settings = "^2"
ant31box = {git = "https://github.com/ant31/ant31box", branch="main"}

[tool.poetry.group.dev.dependencies]
black = "*"
ipython = "*"
pylint = "*"
pyre-check = "*"
types-requests = "*"
isort = "*"
pytest = "*"
coverage = "*"
pytest-cov = "*"
bumpversion = "*"
pytest-ordering = "*"
pytest-asyncio = "*"
pyright = "*"
ruff = "^0.9.2"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.isort]
atomic = true
profile = "black"
line_length = 88
skip_gitignore = true

[tool.pytest]
testpaths = "tests/"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
